Move key material geneerating helpers to tests
